Issue #3177673 by chmez: Fix changing identifier of comment wrapper after adding a first new comment

+++ modules/custom/social_ajax_comments/src/Controller/AjaxCommentsController.php
@@ -15,6 +15,20 @@ use Drupal\ajax_comments\Controller\AjaxCommentsController as ContribController;
 class AjaxCommentsController extends ContribController {
 
   /**
+   * The number of errors.
+   *
+   * @var int|null
+   */
+  protected $errors = NULL;
+
+  /**
+   * TRUE if temporary storage should be cleared.
+   *
+   * @var bool
+   */
+  protected $clearTempStore = TRUE;
+
+  /**
    * Cancel handler for the cancel form.
    *
    * @param \Symfony\Component\HttpFoundation\Request $request
@@ -33,7 +47,7 @@ class AjaxCommentsController extends ContribController {
     $response = new AjaxResponse();
 
     // Get the selectors.
-    $selectors = $this->tempStore->getSelectors($request, $overwrite = TRUE);
+    $selectors = $this->tempStore->getSelectors($request, TRUE);
     $wrapper_html_id = $selectors['wrapper_html_id'];
     $form_html_id = $selectors['form_html_id'];
 
@@ -97,7 +111,7 @@ class AjaxCommentsController extends ContribController {
     // Store the selectors from the incoming request, if applicable.
     // If the selectors are not in the request, the stored ones will
     // not be overwritten.
-    $this->tempStore->getSelectors($request, $overwrite = TRUE);
+    $this->tempStore->getSelectors($request, TRUE);
 
     // Check the user's access to reply.
     // The user should not have made it this far without proper permission,
@@ -126,7 +140,7 @@ class AjaxCommentsController extends ContribController {
     $form = $this->entityFormBuilder()->getForm($comment);
 
     // Check for errors.
-    if (!($errors = count($this->messenger()->messagesByType('error')))) {
+    if (!($this->errors = count($this->messenger()->messagesByType('error')))) {
       // If there are no errors, set the ajax-updated
       // selector value for the form.
       $this->tempStore->setSelector('form_html_id', $form['#attributes']['id']);
@@ -142,7 +156,7 @@ class AjaxCommentsController extends ContribController {
     }
     else {
       // Retrieve the selector values for use in building the response.
-      $selectors = $this->tempStore->getSelectors($request, $overwrite = TRUE);
+      $selectors = $this->tempStore->getSelectors($request, TRUE);
       $wrapper_html_id = $selectors['wrapper_html_id'];
 
       // If there are errors, remove old messages.
@@ -156,29 +170,33 @@ class AjaxCommentsController extends ContribController {
       $cid = $this->tempStore->getCid();
 
       // Try to insert the message above the new comment.
-      if (!empty($cid) && !$errors && \Drupal::currentUser()->hasPermission('skip comment approval')) {
-        $selector = static::getCommentSelectorPrefix() . $cid;
-        $response = $this->addMessages(
-          $request,
-          $response,
-          $selector,
-          'before'
-        );
+      if (
+        !empty($cid) &&
+        !$this->errors &&
+        $this->currentUser()->hasPermission('skip comment approval')
+      ) {
+        $selector = $cid;
+        $position = 'before';
       }
       // If the new comment is not to be shown immediately, or if there are
       // errors, insert the message directly below the parent comment.
       else {
-        $response = $this->addMessages(
-          $request,
-          $response,
-          static::getCommentSelectorPrefix() . $comment->get('pid')->target_id,
-          'after'
-        );
+        $selector = $comment->get('pid')->target_id;
+        $position = 'after';
       }
+
+      $response = $this->addMessages(
+        $request,
+        $response,
+        static::getCommentSelectorPrefix() . $selector,
+        $position
+      );
     }
 
     // Clear out the tempStore variables.
-    $this->tempStore->deleteAll();
+    if ($this->clearTempStore) {
+      $this->tempStore->deleteAll();
+    }
 
     // Remove the libraries from the response, otherwise when
     // core/misc/drupal.js is reinserted into the DOM, the following line of
